Condstore support for MobileMe
Summary:
Briefly, there was a couple bugs with the mailsync code:

1. Some backends need to receive a condstore-enabling message to activate condstore. This is done using a "STATUS CONDSTORE" command.

2. MobileMe sends HIGHESTMODSEQ values bigger than Integer. I changed the db field to be a BigInteger instead. I made the same change to UIDVALIDITY because it's probably the same.

Test Plan: Ran a sync with MobileMe, added a gmail account, checked that condstore worked for both.

Reviewers: emfree, charles

Reviewed By: charles

Projects: #inbox-sync

Maniphest Tasks: T311

Differential Revision: https://review.inboxapp.com/D352

+++ inbox/models/backends/imap.py
@@ -138,7 +138,7 @@ class ImapFolderInfo(MailSyncBase):
                           backref=backref('imapfolderinfo',
                                           primaryjoin=backrefjoin),
                           primaryjoin=primaryjoin)
-    uidvalidity = Column(Integer, nullable=False)
+    uidvalidity = Column(BigInteger, nullable=False)
     # Invariant: the local datastore for this folder has always incorporated
     # remote changes up to _at least_ this modseq (we can't guarantee that we
     # haven't incorporated later changes too, since IMAP doesn't provide a true
@@ -146,7 +146,7 @@ class ImapFolderInfo(MailSyncBase):
     #
     # Note that some IMAP providers do not support the CONDSTORE extension, and
     # therefore will not use this field.
-    highestmodseq = Column(Integer, nullable=True)
+    highestmodseq = Column(BigInteger, nullable=True)
 
     __table_args__ = (UniqueConstraint('account_id', 'folder_id'),)
